pyproject.toml
[project]
name = "rag-anything"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "graspologic>=3.4.1",
    "huggingface-hub>=0.33.0",
    "lightrag-hku>=1.3.9",
    "mineru[core]>=2.0.6",
    "nano-vectordb>=0.0.4.3",
    "pillow>=10.0.0",
]
